move ru jobs to different file
