+++ tgui/packages/tgui/bandastation/ru_jobs.tsx
@@ -1,0 +1,41 @@
+export const JOBS_RU = {
+  Assistant: 'Гражданский',
+  Prisoner: 'Заключенный',
+  Captain: 'Капитан',
+  'Head of Personnel': 'Глава персонала',
+  'Head of Security': 'Глава службы безопасности',
+  'Research Director': 'Директор исследований',
+  'Chief Engineer': 'Главный инженер',
+  'Chief Medical Officer': 'Главный врач',
+  AI: 'ИИ',
+  Cyborg: 'Киборг',
+  'Personal AI': 'Персональный ИИ',
+  'Human AI': 'Большой брат',
+  Warden: 'Смотритель',
+  Detective: 'Детектив',
+  'Security Officer': 'Офицер',
+  'Station Engineer': 'Станционный инженер',
+  'Atmospheric Technician': 'Атмосферный техник',
+  Coroner: 'Коронер',
+  'Medical Doctor': 'Врач',
+  Paramedic: 'Парамедик',
+  Chemist: 'Химик',
+  Scientist: 'Ученый',
+  Roboticist: 'Робототехник',
+  Geneticist: 'Генетик',
+  Quartermaster: 'Квартирмейстер',
+  'Cargo Technician': 'Грузчик',
+  'Shaft Miner': 'Шахтер',
+  Bitrunner: 'Битраннер',
+  Bartender: 'Бармен',
+  Botanist: 'Ботаник',
+  Cook: 'Повар',
+  Chef: 'Шеф',
+  Janitor: 'Уборщик',
+  Clown: 'Клоун',
+  Mime: 'Мим',
+  Curator: 'Куратор',
+  Lawyer: 'Адвокат',
+  Chaplain: 'Священник',
+  Psychologist: 'Психолог',
+};
