+++ tgui/packages/tgui/interfaces/PreferencesMenu/JobsPage.tsx
@@ -3,6 +3,7 @@ import { classes } from 'common/react';
 import { PropsWithChildren, ReactNode } from 'react';
 
 import { useBackend } from '../../backend';
+import { JOBS_RU } from '../../bandastation/ru_jobs'; // BANDASTATION EDIT
 import { Box, Button, Dropdown, Stack, Tooltip } from '../../components';
 import {
   createSetPreference,
@@ -406,48 +407,4 @@ export const JobsPage = () => {
       </Stack>
     </>
   );
-};
-
-// BANDASTATION EDIT START
-
-export const JOBS_RU = {
-  Assistant: 'Гражданский',
-  Prisoner: 'Заключенный',
-  Captain: 'Капитан',
-  'Head of Personnel': 'Глава персонала',
-  'Head of Security': 'Глава службы безопасности',
-  'Research Director': 'Директор исследований',
-  'Chief Engineer': 'Главный инженер',
-  'Chief Medical Officer': 'Главный врач',
-  AI: 'ИИ',
-  Cyborg: 'Киборг',
-  'Personal AI': 'Персональный ИИ',
-  'Human AI': 'Большой брат',
-  Warden: 'Смотритель',
-  Detective: 'Детектив',
-  'Security Officer': 'Офицер',
-  'Station Engineer': 'Станционный инженер',
-  'Atmospheric Technician': 'Атмосферный техник',
-  Coroner: 'Коронер',
-  'Medical Doctor': 'Врач',
-  Paramedic: 'Парамедик',
-  Chemist: 'Химик',
-  Scientist: 'Ученый',
-  Roboticist: 'Робототехник',
-  Geneticist: 'Генетик',
-  Quartermaster: 'Квартирмейстер',
-  'Cargo Technician': 'Грузчик',
-  'Shaft Miner': 'Шахтер',
-  Bitrunner: 'Битраннер',
-  Bartender: 'Бармен',
-  Botanist: 'Ботаник',
-  Cook: 'Повар',
-  Chef: 'Шеф',
-  Janitor: 'Уборщик',
-  Clown: 'Клоун',
-  Mime: 'Мим',
-  Curator: 'Куратор',
-  Lawyer: 'Адвокат',
-  Chaplain: 'Священник',
-  Psychologist: 'Психолог',
 };
